[Fix] #1312 Streak logic duplication resolved - core logic abstracted to src/lib/streak.ts.
