+++ src/app/api/leaderboard/route.ts
@@ -1,6 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { supabaseAdmin } from "@/lib/supabase";
-import { dateDiffDays, toDateStr } from "@/lib/dateUtils";
+import { toDateStr } from "@/lib/dateUtils";
+import { calculateCurrentStreak } from "@/lib/streak";
 import {
   cacheGet,
   cacheSet,
@@ -157,29 +158,6 @@ async function fetchGitHubJson<T>(path: string): Promise<T | null> {
   return (await res.json()) as T;
 }
 
-function calculateCurrentStreak(commitDates: string[]): number {
-  const days = Array.from(new Set(commitDates.map((date) => date.slice(0, 10)))).sort();
-  if (days.length === 0) {
-    return 0;
-  }
-
-  let runLength = 1;
-  const runs: { end: string; length: number }[] = [];
-  for (let i = 1; i < days.length; i += 1) {
-    if (dateDiffDays(days[i - 1], days[i]) === 1) {
-      runLength += 1;
-    } else {
-      runs.push({ end: days[i - 1], length: runLength });
-      runLength = 1;
-    }
-  }
-  runs.push({ end: days[days.length - 1], length: runLength });
-
-  const today = toDateStr(new Date());
-  const yesterday = toDateStr(new Date(Date.now() - 86400000));
-  const latest = runs[runs.length - 1];
-  return latest.end === today || latest.end === yesterday ? latest.length : 0;
-}
 
 async function fetchCommitStats(username: string, since: string) {
   const query = new URLSearchParams({
